% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alpha_posterior}
\alias{alpha_posterior}
\title{Obtain posterior samples from the conditional extremes alpha parameter}
\usage{
alpha_posterior(fitted.mod)
}
\arguments{
\item{fitted.mod}{Object returned from the function fit_GL.}
}
\value{
Matrix of posterior samples from alpha_{2|1} and alpha_{2|1}
}
\description{
Obtain posterior samples from the conditional extremes alpha parameter
}
