% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eta_posterior}
\alias{eta_posterior}
\title{Obtain posterior samples from the coefficient of residual tail dependence eta}
\usage{
eta_posterior(fitted.mod)
}
\arguments{
\item{fitted.mod}{Object returned from the function fit_GL.}
}
\value{
Vector of posterior samples from eta.
}
\description{
Obtain posterior samples from the coefficient of residual tail dependence eta
}
