% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_GL}
\alias{fit_GL}
\title{Obtain realisations from the posterior distribution of the sets G and L}
\usage{
fit_GL(fitted.Qq, config)
}
\arguments{
\item{fitted.Qq}{Object returned by the function fit_Qq.}

\item{config}{Object returned by the function set.configs.}
}
\value{
A list of posterior realisations of G and L and other useful parameters.
}
\description{
Obtain realisations from the posterior distribution of the sets G and L
}
