% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_Qq}
\alias{fit_Qq}
\title{Obtain realisations from the posterior distribution of the quantile set Q_q}
\usage{
fit_Qq(X, options, config, return_fitted_obj = FALSE)
}
\arguments{
\item{X}{n by d data matrix.}

\item{options}{Object returned by the function set.options.}

\item{config}{Object returned by the function set.configs.}

\item{return_fitted_obj}{Boolean: TRUE to return fitted inlabru quantile regression object, FALSE otherwise.}
}
\value{
A list of posterior realisations of Q_q and other useful parameters.
}
\description{
Obtain realisations from the posterior distribution of the quantile set Q_q
}
