% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_2d.R
\name{get_P_W}
\alias{get_P_W}
\title{Calculate the probability that W falls in an interval of the 1-sphere}
\usage{
get_P_W(log.L, mesh, limits.W_B)
}
\arguments{
\item{log.L}{natural logarithm of the unnormalised density of angles W.}

\item{mesh}{INLA object for the cyclic mesh on the 1-sphere.}

\item{limits.W_B}{vector of lower and upper limits of angles (polar) for sampling region.}
}
\value{
probability that W falls in the range limits.W_B.
}
\description{
Calculate the probability that W falls in an interval of the 1-sphere
}
