% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hypersphere_volume}
\alias{hypersphere_volume}
\title{Obtain the volume of a hypersphere}
\usage{
hypersphere_volume(r, d)
}
\arguments{
\item{r}{radius of the sphere.}

\item{d}{dimension of the hypersphere S^{d-1}.}
}
\value{
volume of the hypersphere.
}
\description{
Obtain the volume of a hypersphere
}
