% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{K.envelope}
\alias{K.envelope}
\title{Compute the (1-sig)\% simultaneous envelope for the K function}
\usage{
K.envelope(
  n,
  d,
  subset,
  M = 500,
  sig = 0.95,
  s = seq(0, 2, len = 30),
  phi = seq(0, pi, len = 30)
)
}
\arguments{
\item{n}{number of samples inside the unit hypersphere.}

\item{d}{dimension of the hypersphere S^{d-1}.}

\item{subset}{geometry of the subset on which to count sets, "ball" or "sph cone".}

\item{M}{number of montecarlo samples from a uniform sample on S^{d-1}.}

\item{sig}{credibility level for the envelope.}

\item{s}{sequence of radii for the empirical K function to be estimated (if subset=="ball").}

\item{phi}{sequence of angles for the empirical K function to be estimated (if subset=="sph cone").}
}
\value{
(1-sig)\% simultaneous envelope for the K function.
}
\description{
Compute the (1-sig)\% simultaneous envelope for the K function
}
