% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{K_hat}
\alias{K_hat}
\title{Compute the empirical K function for the fitted model}
\usage{
K_hat(thinned_U, subset, s = seq(0, 2, len = 30), phi = seq(0, pi, len = 30))
}
\arguments{
\item{thinned_U}{thinned matrix of observations inside the unit ball B_1(0) or on S^(d-1).}

\item{subset}{geometry of the subset on which to count sets, "ball" or "sph cone".}

\item{s}{sequence of radii for the empirical K function to be estimated (if subset=="ball").}

\item{phi}{sequence of angles for the empirical K function to be estimated (if subset=="sph cone").}
}
\value{
K function value for the s or phi values.
}
\description{
Compute the empirical K function for the fitted model
}
