% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Khat}
\alias{Khat}
\title{Compute the empirical K function inside the unit ball}
\usage{
Khat(X, s = seq(0, 2, len = 50))
}
\arguments{
\item{X}{matrix of observations in the unit hypersphere.}

\item{s}{sequence of radii for the empirical K function to be estimated.}
}
\value{
K function value for different radii values.
}
\description{
Compute the empirical K function inside the unit ball
}
