% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Khat_B}
\alias{Khat_B}
\title{Compute the empirical K function inside the unit ball via ball subsets}
\usage{
Khat_B(X, s = seq(0, 2, len = 30))
}
\arguments{
\item{X}{matrix of observations inside the unit ball B_1(0).}

\item{s}{sequence of radii for the empirical K function to be estimated.}
}
\value{
K function value for different radii values.
}
\description{
Compute the empirical K function inside the unit ball via ball subsets
}
