% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_G}
\alias{plot_G}
\title{Plot the posterior mean and (1-alpha)-predictive intervals for G}
\usage{
plot_G(
  fitted.mod,
  alpha = 0.05,
  conf = "sim",
  surface3d = "mean",
  surf.col = "grey",
  cex.pts = 0.4,
  cex.axis = 1.4,
  xlim = c(0, 0),
  ylim = c(0, 0),
  xlab = expression(X[1]),
  ylab = expression(X[2]),
  zlab = expression(X[3]),
  by = 2
)
}
\arguments{
\item{fitted.mod}{Object returned by the function fit_GL.}

\item{alpha}{Value in (0,1) for the (1-alpha)-simultaneous predictive interval of G.}

\item{conf}{Choose between simultaneous ("sim") and marginal ("marg") predictive intervals.}

\item{surface3d}{Surface (3d) to plot between the mean ("mean"), and simultaneous predictive interval bounds ("lower" or "upper").}

\item{surf.col}{Color of surface (3d).}

\item{cex.pts}{Size of points representing exceedances of Q_q.}

\item{cex.axis}{Size of axes' labels.}

\item{xlim}{Plotting range on x-axis.}

\item{ylim}{Plotting range on y-axis.}

\item{xlab}{Label of the x-axis.}

\item{ylab}{Label of the y-axis.}

\item{zlab}{Label of the z-axis.}

\item{by}{Length of intervals between axis ticks and displayed values.}
}
\value{
None.
}
\description{
Plot the posterior mean and (1-alpha)-predictive intervals for G
}
