% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_Qq}
\alias{plot_Qq}
\title{Plot the posterior mean and (1-alpha)-predictive intervals for Q_q}
\usage{
plot_Qq(
  fitted.Qq,
  alpha = 0.05,
  conf = "sim",
  surface3d = "mean",
  cex.pts = 0.4,
  cex.axis = 1.4,
  xlim = c(0, 0),
  ylim = c(0, 0),
  by = 2
)
}
\arguments{
\item{fitted.Qq}{Object returned by the function fit_Qq.}

\item{alpha}{Value in (0,1) for the (1-alpha)-simultaneous predictive interval of Q_q.}

\item{conf}{Choose between simultaneous ("sim") and marginal ("marg") predictive intervals.}

\item{surface3d}{Surface to plot between the mean ("mean"), and simultaneous predictive interval bounds ("lower" or "upper").}

\item{cex.pts}{Size of points representing exceedances of Q_q.}

\item{cex.axis}{Size of axes' labels.}

\item{xlim}{Plotting range on x-axis.}

\item{ylim}{Plotting range on y-axis.}

\item{by}{Length of intervals between axis ticks and displayed values.}
}
\value{
None.
}
\description{
Plot the posterior mean and (1-alpha)-predictive intervals for Q_q
}
