% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_W}
\alias{plot_W}
\title{Plot the posterior mean and (1-alpha)-predictive intervals for W}
\usage{
plot_W(
  fitted.mod,
  alpha = 0.05,
  conf = "sim",
  f_W.lim = 0,
  main = "",
  mid.gap = 0.1,
  txt.gap = 0.02,
  cex.txt = 1.4,
  exconly = FALSE
)
}
\arguments{
\item{fitted.mod}{Object returned by the function fit_GL.}

\item{alpha}{Value in (0,1) for the (1-alpha)-simultaneous predictive interval of W.}

\item{conf}{Choose between simultaneous ("sim") and marginal ("marg") predictive intervals.}

\item{f_W.lim}{Plotting range of the image of the density f_W.}

\item{main}{Title of figure.}

\item{mid.gap}{Radius of the middle circle of zero density.}

\item{txt.gap}{Gap between angle marks and axes' lines.}

\item{cex.txt}{Size of angle marks.}

\item{exconly}{Boolean: TRUE if circular histogram should be based only on exceedances, FALSE for all angles.}
}
\value{
None.
}
\description{
Plot the posterior mean and (1-alpha)-predictive intervals for W
}
