% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_X_t}
\alias{plot_X_t}
\title{Plot the boundary of the return set X_t and its simultaneous predictive intervals}
\usage{
plot_X_t(
  fitted.mod,
  list_ret_sets,
  plt = "boundary",
  xylim = c(0, 0),
  xyzlim = c(0, 0),
  xlab = expression(X[1]),
  ylab = expression(X[2]),
  zlab = expression(X[3]),
  by = 2,
  cex.pts = 0.4,
  cex.axis = 1.4
)
}
\arguments{
\item{fitted.mod}{Object returned by the function fit_GL.}

\item{list_ret_sets}{Object returned by the function return_set.}

\item{plt}{Plot return level-set ("set") or boundary ("boundary").}

\item{xylim}{Plotting limits of x and y axes.}

\item{xyzlim}{Plotting limits of x, y, and y axes.}

\item{xlab}{Label of x-axis.}

\item{ylab}{Label of y-axis.}

\item{zlab}{Label of z-axis.}

\item{by}{Length of intervals between axis ticks and displayed values.}

\item{cex.pts}{Size of observed data points.}

\item{cex.axis}{Size of text on axes.}
}
\value{
None
}
\description{
Plot the boundary of the return set X_t and its simultaneous predictive intervals
}
