% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prob_estimation}
\alias{prob_estimation}
\title{Probability estimation}
\usage{
prob_estimation(fitted.mod, post.sample, x_B, y_B)
}
\arguments{
\item{fitted.mod}{Object returned from the function fit_GL.}

\item{post.sample}{Object returned from the function sample_QGW_posterior.}

\item{x_B}{x-component of the rectangular set B of interest.}

\item{y_B}{y-component of the rectangular set B of interest.}
}
\value{
Vector of posterior samples for the probability that a new observation falls in B.
}
\description{
Probability estimation
}
