% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{return_set}
\alias{return_set}
\title{Obtain realisations from the radial function of the return level-set X_t.}
\usage{
return_set(fitted.mod, t = NA, q.prime = NA)
}
\arguments{
\item{fitted.mod}{Object returned from the function fit_GL.}

\item{t}{Vector of return periods t.prime <= 1/(1-q) for the return level-sets X_{t.prime}.}

\item{q.prime}{Vector of probabilities of the canonical return level-set level-sets X_{1/(1-q.prime)}.}
}
\value{
Matrix of realisations from the radial function of the return level-set X_t.
}
\description{
Obtain realisations from the radial function of the return level-set X_t.
}
