% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_2d.R
\name{return_set_2d_isotropic}
\alias{return_set_2d_isotropic}
\title{Title}
\usage{
return_set_2d_isotropic(
  fitted.mod,
  alpha = 0.05,
  conf = "sim",
  t,
  n.MC.samp = 10000
)
}
\arguments{
\item{n.MC.samp}{Number of integrating directions if set = "isotropic".}
}
\description{
Title
}
