% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_QGW_posterior}
\alias{sample_QGW_posterior}
\title{Sample angles from W and obtain g and G evaluated at these angles.}
\usage{
sample_QGW_posterior(fitted.mod, N.w, S_B = NA, transf.G = FALSE)
}
\arguments{
\item{fitted.mod}{Object returned from the function fit_GL.}

\item{N.w}{Number of angles to sample.}

\item{S_B}{Subset of the 1-sphere S from -pi to pi. NA for S.}

\item{transf.G}{Boolean, use transf.G if TRUE, G otherwise.}
}
\description{
Sample angles from W and obtain g and G evaluated at these angles.
}
