% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_beta}
\alias{select_beta}
\title{Obtain un-normalised log-posterior of beta (1/(1-beta)-power transformation of the radius R)}
\usage{
select_beta(X, q, beta.seq, config)
}
\arguments{
\item{X}{matrix (n by d) of observations.}

\item{q}{quantile of the quantile set Q_q to fit.}

\item{beta.seq}{sequence of beta values between 0 and 1.}

\item{config}{save configurations.}
}
\description{
Obtain un-normalised log-posterior of beta (1/(1-beta)-power transformation of the radius R)
}
