% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set.configs}
\alias{set.configs}
\title{Set configurations used by functions fit_Qq and fit_GL}
\usage{
set.configs(
  save.path = "",
  file.nm = paste0(options$data.nm, "_", options$excess.dist.fam, "_", options$W.model,
    "_", options$W.data),
  save = FALSE,
  progress = FALSE
)
}
\arguments{
\item{save.path}{Path to folder where .RData objects are to be saved.}

\item{file.nm}{Name of the .RData file.}

\item{save}{Boolean: TRUE to save .RData files, FALSE otherwise.}

\item{progress}{Boolean: TRUE to save progression in a .txt file, FALSE otherwise.}
}
\value{
A list of configurations.
}
\description{
Set configurations used by functions fit_Qq and fit_GL
}
\examples{
config <- set.configs(save.path = "path/to/folder/",
                      file.nm   = "file_name", # default based on options if no argument given
                      save      = TRUE,
                      progress  = TRUE)
}
