% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set.options}
\alias{set.options}
\title{Set options used by functions fit_Qq and fit_GL}
\usage{
set.options(
  X,
  excess.dist.fam,
  W.model,
  W.data,
  use.mean.Qq,
  q,
  alpha,
  N.Qq,
  N.GW,
  QR.prior.range = c(1, 0.999),
  QR.prior.sigma = c(1, 0.8),
  zeta.prior.range = c(0.5, 0.999),
  zeta.prior.sigma = c(5, 0.8),
  phi.prior.range = c(0.5, 0.999),
  phi.prior.sigma = c(5, 0.8),
  mesh.knots.2d = seq(-pi, pi, by = pi/400),
  mesh.res.3d = 20,
  seed = 0L
)
}
\arguments{
\item{X}{n by d observations matrix.}

\item{excess.dist.fam}{Exponential ("E") or generalised Pareto ("GP") excesses.}

\item{W.model}{"M1", "M2", or "M3", as indicated in Papastathopoulos et al. (2023).}

\item{W.data}{Angles to include in likelihood: all angles ("AllW") or only those of exceedances of Q_q ("ExcOnly").}

\item{use.mean.Qq}{Boolean: use mean of Qq to fit G and L if TRUE, use samples from Qq otherwise.}

\item{q}{Quantile of the quantile set Q_q to fit.}

\item{alpha}{Quantile of the generalised Pareto to use in fitting parameterisation, alpha in (0,1).}

\item{N.Qq}{Number of quantile sets to sample from the posterior distribution of Q_q. Not used if use.mean.Qq is TRUE.}

\item{N.GW}{Number of G and W sets to sample from the posterior distribution of G and W.}

\item{QR.prior.range}{A length 2 vector, with (range0,Prange) specifying that P(rho < rho_0)=p_rho, where rho is the spatial range of the random field Q_q. If Prange is NA, then range0 is used as a fixed range value.}

\item{QR.prior.sigma}{A length 2 vector, with (sigma0,Psigma) specifying that P(sigma > sigma_0)=p_sigma, where sigma is the marginal standard deviation of the field Q_q. If Psigma is NA, then sigma0 is used as a fixed range value.}

\item{zeta.prior.range}{A length 2 vector, with (range0,Prange) specifying that P(rho < rho_0)=p_rho, where rho is the spatial range of the random field G. If Prange is NA, then range0 is used as a fixed range value.}

\item{zeta.prior.sigma}{A length 2 vector, with (sigma0,Psigma) specifying that P(σ > σ_0)=p_σ, where σ is the marginal standard deviation of the field B. If Psigma is NA, then sigma0 is used as a fixed range value.}

\item{phi.prior.range}{A length 2 vector, with (range0,Prange) specifying that P(rho < rho_0)=p_rho, where rho is the spatial range of the random field B. If Prange is NA, then range0 is used as a fixed range value.}

\item{phi.prior.sigma}{A length 2 vector, with (sigma0,Psigma) specifying that P(σ > σ_0)=p_σ, where σ is the marginal standard deviation of the field B. If Psigma is NA, then sigma0 is used as a fixed range value.}

\item{mesh.knots.2d}{Sequence of knots in a subset of of the interval -pi to pi.}

\item{mesh.res.3d}{Subdivision resolution for a semi-regular spherical triangulation with equidistant points along equidistant latitude bands.}

\item{seed}{Random number generator seed passed on to INLA::inla.posterior.sample.}
}
\value{
A list of options.
}
\description{
Set options used by functions fit_Qq and fit_GL
}
\examples{
X <- rmutil::qlaplace(cbind(runif(1000),runif(1000),runif(1000)))
options <- set.options(X                = X,
                       excess.dist.fam  = "E",
                       W.data           = "ExcOnly",
                       W.model          = "M3",
                       use.mean.Qq      = "FALSE",
                       q                = 0.9,
                       alpha            = 0.5,
                       N.Qq             = 20,
                       N.GW             = 50,
                       QR.prior.range   = c(1, 0.999),
                       QR.prior.sigma   = c(1, 0.8),
                       zeta.prior.range = c(0.5, 0.999),
                       zeta.prior.sigma = c(5, 0.8),
                       phi.prior.range  = c(0.5, 0.999),
                       phi.prior.sigma  = c(5, 0.8),
                       mesh.knots.2d    = seq(-pi,pi,by=pi/400),
                       mesh.res.3d      = 20,
                       seed             = 0L)
}
