% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{thin_U_on_ball}
\alias{thin_U_on_ball}
\title{Thin the transformed point process to have same number of observations}
\usage{
thin_U_on_ball(U_on_Ball)
}
\arguments{
\item{U_on_Ball}{Object returned by the X_to_uniform_on_Ball function.}
}
\description{
Thin the transformed point process to have same number of observations
}
