% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{toLaplaceMargins}
\alias{toLaplaceMargins}
\title{Transform matrix of observed data to matrix with Laplace marginal distributions}
\usage{
toLaplaceMargins(X, q)
}
\arguments{
\item{X}{n by d data matrix.}

\item{q}{2 by d matrix. The top/bottom row denote the upper/lower quantiles above/below which to fit a GPd. NA for empirical.}
}
\value{
A list including the transformed data and the marginal transformation parameters.
}
\description{
Transform matrix of observed data to matrix with Laplace marginal distributions
}
\examples{
n <- 500
X <- cbind(rnorm(n,0,1),rnorm(n,0,3),rnorm(n,0,5),rnorm(n,0,7))
q <- cbind(c(0.95,0.05),c(NA,0.05),c(0.95,NA),c(NA,NA))
LapTransf <- toLaplaceMargins(X,q)
}
