% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{toOriginalMargins}
\alias{toOriginalMargins}
\title{Transform matrix of data with Laplace marginal distributions to matrix with original marginal distributions}
\usage{
toOriginalMargins(X.L, LapTransf)
}
\arguments{
\item{X.L}{n by d matrix of observations in Laplace margins, same dimensions as LapTransf$X.}

\item{LapTransf}{Object returned by the function toLaplaceMargins.}
}
\value{
An n by d matrix of observations in original marginal distributions.
}
\description{
Transform matrix of data with Laplace marginal distributions to matrix with original marginal distributions
}
\examples{
n <- 500
X <- rnorm(n,0,1)
q <- cbind(c(0.95,0.05))
LapTransf <- toLaplaceMargins(X,q)
X.L <- rmutil::qlaplace(runif(100))
X.Orig <- toOriginalMargins(X.L,LapTransf)
}
