% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{X_to_uniform_on_Ball}
\alias{X_to_uniform_on_Ball}
\title{Transform the observations X to a uniform inside the unit hypersphere}
\usage{
X_to_uniform_on_Ball(fitted.mod)
}
\arguments{
\item{fitted.mod}{Object returned from the function fit_GL.}
}
\value{
list of samples from the transformed point process.
}
\description{
Transform the observations X to a uniform inside the unit hypersphere
}
