% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Khat_W}
\alias{Khat_W}
\title{Compute the empirical K function inside the unit ball via spherical cone subsets}
\usage{
Khat_W(W, phi = seq(0, pi, len = 30))
}
\arguments{
\item{W}{matrix of observations on S^(d-1).}

\item{phi}{sequence of angles of the spherical cone for the empirical K function to be estimated.}
}
\value{
K function value for different angular values.
}
\description{
Compute the empirical K function inside the unit ball via spherical cone subsets
}
