% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{return_set}
\alias{return_set}
\title{Obtain realisations from the radial function of the return level-set X_t.}
\usage{
return_set(
  fitted.mod,
  set = "classic",
  alpha = 0.05,
  conf = "sim",
  t = NA,
  q.prime = NA,
  include.Qq = FALSE,
  LapTransf = NA,
  n.MC.samp = 10000
)
}
\arguments{
\item{fitted.mod}{Object returned from the function fit_GL.}

\item{set}{Type of quantile set, either "classic" or "isotropic".}

\item{alpha}{Value in (0,1) for the (1-alpha)-simultaneous predictive interval of X_t.}

\item{conf}{Choose between simultaneous ("sim") and marginal ("marg") predictive intervals.}

\item{t}{Vector of return periods t.prime <= 1/(1-q) for the return level-sets X_{t.prime}.}

\item{q.prime}{Vector of probabilities of the canonical return level-set level-sets X_{1/(1-q.prime)}.}

\item{include.Qq}{Boolean: Include Q_q in the return sets if TRUE, do not include otherwise.}

\item{LapTransf}{Object returned by the function toLaplaceMargins. If provided, return sets will be plotted in original margins.}

\item{n.MC.samp}{Number of integrating directions if set = "isotropic".}
}
\value{
Matrix of realisations from the radial function of the return level-set X_t.
}
\description{
Obtain realisations from the radial function of the return level-set X_t.
}
