% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{return_set}
\alias{return_set}
\title{Obtain realisations from the radial function of the return level-set X_t.}
\usage{
return_set(fitted.mod, t = NA, q.prime = NA)
}
\arguments{
\item{fitted.mod}{Object returned from the function fit_GL.}

\item{t}{Return period t <= 1/(1-q) for the return level-set X_t.}

\item{q.prime}{Probability of the canonical return level-set level-set X_{1/(1-q)}.}
}
\value{
Matrix of realisations from the radial function of the return level-set X_t.
}
\description{
Obtain realisations from the radial function of the return level-set X_t.
}
